Refatora a gramática separando-a em analisadores léxico, sintático e semântico.
- Cria `analisador_léxico.js` para regras de espaços e literais básicos.
- Cria `analisador_semântico.js` para funções de transformação e operadores.
- Cria `analisador_sintático.js` para a estrutura gramatical.
- Atualiza `0.js` para importar e utilizar os novos módulos.
